Split JS functions
