+++ .pylintrc
@@ -1,2 +1,2 @@
 [MASTER]
-disable=C0114,E0401,C0209
+disable=C0114,E0401,C0209,W0703,R1710,R0902,R0913,E1101
