+++ templates/index.html
@@ -3,11 +3,21 @@
 	<head>
 		<meta charset="utf-8">
 		<title>Plotter 2D</title>
+		<!-- Add global styles -->
 		<link rel="stylesheet" href="{{ url_for('static', filename='css/styles.css') }}">
+		<!-- Add Bootstrap -->
 		<link href="https://cdn.jsdelivr.net/npm/bootstrap@5.1.3/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-1BmE4kWBq78iYhFldvKuhfTAU6auU8tT94WrHftjDbrCEXSU1oBoqyl2QvZ6jIW3" crossorigin="anonymous">
 		<link rel="pencil icon" href="{{ url_for('static', filename='favicon.ico') }}">
+		<!-- Add AJAX -->
 		<script src="https://ajax.googleapis.com/ajax/libs/jquery/3.5.1/jquery.min.js"></script>
+		<!-- Add custom scripts used by the webserver -->
 		<script src="{{ url_for('static', filename='scripts/utils.js') }}"></script>
+		<script src="{{ url_for('static', filename='scripts/navigation.js') }}"></script>
+		<script src="{{ url_for('static', filename='scripts/camera.js') }}"></script>
+		<script src="{{ url_for('static', filename='scripts/plot.js') }}"></script>
+		<script src="{{ url_for('static', filename='scripts/images.js') }}"></script>
+		<script src="{{ url_for('static', filename='scripts/stats.js') }}"></script>
+		<script src="{{ url_for('static', filename='scripts/upload.js') }}"></script>
 	</head>
 	<body>
 		<div class="container d-flex justify-content-center main">
